Reorganize persistence entities
