+++ Sphynx.ServerV2/Persistence/Room/IRoomRepository.cs
@@ -2,10 +2,9 @@
 // See the LICENCE file in the repository root for full licence text.
 
 using Sphynx.Core;
-using Sphynx.ModelV2;
 using Sphynx.ModelV2.Room;
 
-namespace Sphynx.ServerV2.Persistence
+namespace Sphynx.ServerV2.Persistence.Room
 {
     public interface IRoomRepository
     {
@@ -19,5 +18,8 @@ namespace Sphynx.ServerV2.Persistence
         Task<SphynxErrorInfo<IChatRoomInfo?>> GetRoomAsync(SnowflakeId roomId, CancellationToken cancellationToken = default);
 
         Task<SphynxErrorInfo<IChatRoomInfo[]?>> GetRoomsAsync(SnowflakeId[] roomIds, CancellationToken cancellationToken = default);
+
+        Task<SphynxErrorInfo<T?>> GetRoomFieldAsync<T>(SnowflakeId roomId, string fieldName, CancellationToken cancellationToken = default);
+        Task<SphynxErrorCode> UpdateRoomFieldAsync<T>(SnowflakeId roomId, string fieldName, T value, CancellationToken cancellationToken = default);
     }
 }
